[FIX] How the parameters are passed around and change time to be of type long

+++ osu-Replay-Analyzer/osu!/osu_standard.h
@@ -10,7 +10,7 @@
 
 namespace OSUSTANDARD
 {
-	bool isHitobjectAt(std::vector<Hitobject*>& _hitobjects, double _prevTime, double _currTime);
+	bool isHitobjectAt(std::vector<Hitobject*>& _hitobjects, long _prevTime, long _currTime);
 
 	std::vector<std::pair<irr::core::vector2d<double>, double>> getPattern(std::vector<Hitobject*>* _hitobjects, int _num, double _interval, double _time, bool _skipSliders = false);
 	int getButtonState(int _prevKey, int _currKey);
